Fix fallback to custom push notifications.
